refactor(js): split scan helpers
