+++ internal/lang/js/scan_repo.go
@@ -1,0 +1,45 @@
+package js
+
+import (
+	"context"
+	"io/fs"
+
+	"github.com/ben-ranford/lopper/internal/lang/shared"
+)
+
+type scanRepoState struct {
+	parser          *sourceParser
+	repoPath        string
+	result          *ScanResult
+	parseErrorCount int
+	parseErrorFiles []string
+}
+
+func scanRepoEntry(ctx context.Context, state *scanRepoState, path string, entry fs.DirEntry) error {
+	if entry.IsDir() {
+		if shared.ShouldSkipDir(entry.Name(), skipDirectories) {
+			return fs.SkipDir
+		}
+		return nil
+	}
+	if !isSupportedFile(path) {
+		return nil
+	}
+
+	content, tree, relPath, err := readAndParseFile(ctx, state.parser, state.repoPath, path)
+	if err != nil {
+		return err
+	}
+	if tree.RootNode().HasError() {
+		state.parseErrorCount++
+		appendParseErrorFile(&state.parseErrorFiles, relPath)
+	}
+	state.result.Files = append(state.result.Files, analyzeFile(tree, content, relPath))
+	return nil
+}
+
+func appendParseErrorFile(parseErrorFiles *[]string, relPath string) {
+	if len(*parseErrorFiles) < 5 {
+		*parseErrorFiles = append(*parseErrorFiles, relPath)
+	}
+}
